refactor(Parser): extract language independent bits to GenericParser class
